requirements.txt
flask
flask_sqlalchemy
flask_migrate
flask_cors
flask_jwt_extended
python-dotenv
psycopg2-binary
flasgger
